Created User Creation functionality

+++ base/views.py
@@ -1,9 +1,12 @@
+from multiprocessing import context
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
 from .models import Room, Topic
 from django.contrib.auth.models import User, Group
 from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 from .forms import RoomForm
 from django.db.models import Q
 from django.template import loader
@@ -17,7 +20,7 @@ from django.template import loader
 # ]
 
 def loginPage(request):
-
+    page = 'login'
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -36,8 +39,9 @@ def loginPage(request):
             print(messages.error)
             messages.error(request, 'Username or password is incorrect')
 
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
+
 
 def logoutUser(request):
     print(request)
@@ -45,7 +49,20 @@ def logoutUser(request):
     print(request)
     return redirect(home)
 
+def registerPage(request):
+    page = 'register'
+    form = UserCreationForm()
+    context = {'page': page, 'form': form}
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            form.save()
 
+    return render(request, 'base/login_register.html', context)
+
+
+def notAuthorized(request):
+    return render(request, 'base/not_authorized.html')
 
 def home(request):
     search = request.GET.get('search_for') if request.GET.get('search_for') != None else ''
@@ -58,7 +75,7 @@ def home(request):
         )
     topics = Topic.objects.all()
     users = User.objects.all()
-    print(request.POST)
+    print(request.user.is_superuser)
     roomCount = len(rooms)
     context = {'rooms': rooms, 'topics': topics, 'users': users, 'search': search, 'roomCount': roomCount}
     return render(request, 'base/home.html', context)
@@ -72,19 +89,27 @@ def room(request, pk): #To populate URI with room id
     context = {'room': rooms}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
         form = RoomForm(request.POST)
+
         if form.is_valid():
             form.save()
             return redirect('home')
     context = {'forms': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
+
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return redirect('notAuthorized')
+
     if request.method =='POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -93,9 +118,13 @@ def updateRoom(request, pk):
     context = {'forms': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk) 
-    print(request.META)
+
+    if request.user != room.host:
+        return redirect('notAuthorized')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
